Move several components from Publish/Reader to commonPublish
Deliberately doing nothing else in this commit to help git history.
