+++ src/BloomBrowserUI/publish/commonPublish/CoverColorGroup.tsx
